# src/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: App Reviews Analyzer
  description: API for App Reviews Analyser

components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
security:
  - BasicAuth: []

paths:
  /compute/scoreDistribution:
    post:
      description: Returns a bar chart as PNG image to show the review's score distribution between 2 period of time
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                    description: Date to compare score distribution (if empty, no comparaison)
                    type: string
      responses:
        '200':
          description: Returns a bar chart as PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
  /compute/means:
    post:
      description: Returns a graphic as PNG image
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                timeDelta:
                  description: Define the duration in days on which it computes cumulative results (default = 30)
                  type: integer
                ignore:
                  description: Indicates the number of first reviews to skip (default = 0)
                  type: integer
      responses:
        '200':
          description: Returns a graphic as PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
  /wordcloud/computeWords:
    post:
      description: Returns a 2 set of vocabulary and their frequencies 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alpha:
                  description: Parameters to reduce noises in data
                  type: integer
                n:
                  description: Define the number of word per token vocabulary (group of consecutive word)
                  type: integer
                lang:
                  description: Language of analysed reviews from the following list ["de","en","es","fr","it","ja","pt","zh"]
                  type: string
                score:
                  description: Filter reviews on score (take all if empty or 0)
                  type: integer
                start1:
                  description: First period start date to compare
                  type: string
                end1:
                  description: First period end date to compare
                  type: string
                start2:
                  description: Second period start date to compare
                  type: string
                end2:
                  description: Second period end date to compare
                  type: string


      responses:
        '200':
          description: Returns a JSON object including the 2 set of vocabulary
          content:
            application/json:
              schema:
                type: array
                items:
                  type: array
                  items:
                    type: array
                    items:
                      oneOf:
                      - type: string
                      - type: integer
  /wordcloud/generateImage:
    post:
      description: Returns a wordcloud as PNG image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
      responses:
        '200':
          description: Returns a graphic as PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary

          